Deleted the old Runefile syntax and any code that used it
