+++ crates/codegen/src/code.rs
@@ -575,7 +575,7 @@ mod tests {
     fn rune(doc: &str) -> Rune {
         let doc = Document::parse(doc).unwrap();
         let mut diags = Diagnostics::new();
-        let rune = hotg_rune_syntax::analyse_yaml_runefile(&doc, &mut diags);
+        let rune = hotg_rune_syntax::analyse(&doc, &mut diags);
         assert!(diags.is_empty(), "{:#?}", diags);
 
         rune
