+++ crates/syntax/Cargo.toml
@@ -18,8 +18,6 @@ codespan = { version = "0.11.0", features = ["serialization"] }
 codespan-reporting = "0.11.0"
 indexmap = { version = "1.6.2", features = ["serde-1"] }
 once_cell = "1.7.2"
-pest = "2.1.3"
-pest_derive = "2.1.0"
 regex = "1.5.4"
 serde = { version = "1.0.124", features = ["derive"] }
 serde_yaml = "0.8.17"
