--- crates/syntax/src/yaml/mod.rs
@@ -1,5 +1,0 @@
-mod convert;
-mod types;
-
-pub use types::*;
-pub use convert::document_from_runefile;
